CSS fixes
This is a first step for improving Desktop pages.

Mobile rensponsiveness is still missing. However I need to check with LaSuite design studio about the HTML+CSS they are using. On Docs and Visio it look like rensponsive design is working better. We may be using the wrong boiler plate.